server: Initial dynamic endpoints
This is initial implementation of dynamic endpoints. When a type is
defined, an "read all" endpoint is generated.

For example:

% curl -I -X GET localhost:3000/position
HTTP/1.1 404 Not Found
content-length: 0
date: Wed, 22 Sep 2021 10:17:51 GMT

% cargo run --bin chisel type define cli/examples/person.graphql
Type defined: Person
Type defined: Position

% curl -I -X GET localhost:3000/position
HTTP/1.1 200 OK
content-length: 2
date: Wed, 22 Sep 2021 10:19:06 GMT

The endpoint logic is implemented in ApiService, which is a small
wrapper on top of Hyper library. We might want to replace ApiService
with a more advanced HTTP router or a Radix-tree based matcher:

https://github.com/ibraheemdev/httprouter-rs

https://github.com/ibraheemdev/matchit

But this implementation is sufficient for now.