Run the provided javascript on endpoints
After trying a bunch of alternatives, I don't think there is any way
around the fact that a v8 isolate doesn't want to be moved between or
used from multiple threads.

A JsRuntime owns an isolate, so the only solution I can think of is
what this implements: Use a thread local storage.

This has an interesting implication: We cannot easily provide a way to
hold transient server state, since each request can hit a different
thread. A client that wants that would have to put the information in
a database or cookie as appropriate.

The above is probably fine, since at some point we will be shading
our server, so there is not going to be a single process anyway.