Move type names to another table
Since the current idea for implementing type rename is to use alias
and not rename the table, move the name out of the type table.

With this all aliases are symmetrical. They are just a name to type-id
mapping.

The type table is where we will add information specific to the type
itself: field, table name, properties, etc.