Librarify chiseld
As pointed out by Dejan, we have a implicit dependency to "chiseld" from
the "cli" project tests. However, as it's a binary dependency, "cargo"
won't build it automatically for "cargo test".

As a first step, let's library chiseld so that we can make a dev
dependency from cli to the server. We need to move some more code from
main() to the library for this to work, but let's start with the simple
stuff.

Refs: #42