Docker image for integration tests
For integration tests, we'll want a docker image with databases, libraries, dependencies,
and whatnot.

This adds a very basic docker image (databases are installed, but not configured), and github
actions that is smart enough not to rebuild the base image all the time.

No integration tests are yet run in that image, but that can be hooked up by one of you later.