Support async endpoints
This is important since the data the endpoint is using will be coming
from a database.

To test this in isolation, this patch enables network access and
forwards a fetch request.

I have locally tested it by downloading a fedora iso image and
comparing the sha256. The image is 2.1GB. Downloading it, chiseld used
56.24 seconds in user space and 33.89 seconds in system. The peak
memory allocation was 19 MB and the total memory allocation was 12 GB.

In summary, there is room for improvement, but this is not bad for a
first version and streaming is working.