Replace the rusty_v8 cache hack
Unfortunately it looks like a successful run doesn't fully replace an
old cache, so we have to keep avoiding the old copy.

Instead of deleting the broken files, set a key in the config. With
this we will get a full build the first time and then should have a
working cache.