Copy request bodies from rust to javascript
There are still a lot of FIXMEs in the code, but I think it is better
to get this in an iterate as it at least fixes the existing test and
provides a "simple" example of using deno ops.

This also changes the example to be a plain echo server. I tested
uploading and downloading back an iso image with

$ curl -T Fedora-Workstation-Live-x86_64-35_Beta-1.2.iso -s -S -o out localhost:3000/echo

chiseld used 1.24 seconds in user space and 0.68 seconds is system to
run that.

The peak heap usage was 75 MB and the total heap allocation was 4 GB.