Add a new test
This hits what is the hard problem with the rust to js mapping. When a
js Promise depends on a Rust future to resolve, we have to register
that future with deno or we get the error:

   Promise resolution is still pending but the event loop has already resolved.

In particular, it hits it by JS having a promise for reading the
request body that will only be available when the future created by
hyper is ready. Since rust futures have to be polled and js promises
don't, a bridge is need, which not currently used.

The test passes as current master is not trying to read the request
body. This test just prevents me from adding broken code for doing
that read.