Turn some rust code into javascript
This is much simpler as we don't have to manually create closures.

The code in question was just manipulating JS objects, so I don't
expect any performance differences.

I tested uploading and downloading back an iso image with

$ curl -T Fedora-Workstation-Live-x86_64-35_Beta-1.2.iso -s -S -o out localhost:3000/echo

chiseld used 0.93s seconds in user space and 0.76 seconds is system to
run that.

The peak heap usage was 70 MB and the total heap allocation was 4 GB.