server/query: Fix repeated field names in types
We currently have unique constraint on the "field_name" but this is clearly bogus: different types are allowed to have the same field name.

A proper fix is to define unique constraint on (field_name, field_id)
tuple, but sea-ql does not support that:

https://github.com/SeaQL/sea-query/issues/176

Therefore, drop the unique constraint from "field_name" for now to allow
multiple types to have fields with the same name.

Fixes: #163