Support more native types
GraphQL supports natively String, Float, Int and Boolean.
Add those to the list.

At the moment it kind of works if you happen to pass something
well-behaved because Javascript is Javascript. As we discussed,
Jan will likely tackle validation at a different layer later so now
we're just adding the types to the list of accepted types.