Add back constraint on unique field names
This implements Glauber's suggestion. The issue with trying to use an
index is that we don't have a single table with all the names present
in a type. It is possible to create a view:

CREATE VIEW unique_field_names AS
  SELECT types.type_id, field_names.field_name
    FROM types, fields, field_names
    ON types.type_id = fields.type_id AND field_names.field_id = fields.field_id

But sqlite doesn't support indexes on views and postgresql only
supports them on materialized views, which have to be manually
updated.

With postgresql it might be possible to use function with a CHECK
clause, but that seems to be bending the rules a bit.