Route based on largest prefix
This allows having a single endpoint that handles /user/1 and user/2
for example.

This is a lot simpler than what implemented by other frameworks. See
for example:

* https://expressjs.com/en/5x/api.html#router
* https://rocket.rs/v0.5-rc/guide/requests/

But this has the advantage of not needing to check the endpoints in a
particular order (express) or having a rank heuristic and producing
errors in the case of collisions (rocket).

To be revisited once we have more of the publishing workflow settled.