Don't recreate client if rpc fails
When we are waiting for the server, it can happen that we are able to
connect, but the server is starting up. I don't think it ever happens
with the current implementation, but it probably good practice to
handle it.

Before this patch we would recreate the connection, now we keep it and
just retry the message.