prepare for TPC
In preparation for TPC, move somethings that are global to a run-once
shared initialization state. Those are things like signal handlers, the
RPC server, etc.

Conceptually, each thread should have its own API server, and runtime
state so we don't contend for them (although at first they may be shared
to speed up development). So they are kept, together with the deno
initialization, in the per-thread function.

We will create one thread per executor. At this point, we won't be able
to support more than one, though. More work needs to be done for that.