Add the API service to the Runtime
The API service is still created in the shared state, and then passed to
the executor-specific state.

This is because there is a dependency on the RPC service, that wants to
add routes to the API.

However, everywhere else the RPC service just does runtime().get. As a
preparation step towards moving this to a TPC architecture we will break
the dependency between RPC and API and allow RPC to acquire the API from
the current runtime.

The next step is to add a control plane so that RPC can issue commands
to all executors.