massage the RPC server so that it becomes easier to do TPC
We want to have a single RPC server in the whole system, which means
we will have to pass commands down to the individual executors to inform
them of policies being loaded, endpoints being created, etc.

The way this is done today, those things can fail. That means that we
would now have to deal with a complicated state transition, map reducing
errors, rolling back things that succeeded in some executors but failed
in others, etc.

However because most of our entities come from an external source of
truth, we should be able to split them in two steps: one that does
validation, is global, and can fail, and then the local part can be
a one-way channel that is guaranteed to succeed.

The one exception is the type system. That is a bit more complex because
it is not stateless and it could potentially fail if a type already
exists, or in the future, if a type change in the database fails.

So for types, we will keep a global view of what the types are and apply
validation and type evolution globally. After that is done, we propagate
it to the executors in an infallible way.