remove endpoints that are removed from the filesystem
We don't have an RPC verb to remove an endpoint, so once an endpoint is
removed from the filesystem the route keeps existing which is obviously
not what we want.

This patch implements rpc verbs that remove paths given a regular
expression, and then uses that to delete all routes before chisel apply
adds them back.