factor out database connection
Both the query engine and the metadata store repeat code that deal with
connecting with the database - which is already an indication that we're
better off factoring this out.

However an even more important reason is that in preparation for a TPC
workflow, a problem we currently have becomes apparent:

Databases that are external to chiseld (like postgres), and databases
that are in-process (like sqlite) behave very differently when it comes
to how connections are handled.

For external databases, if we have a new executor thread we will want
that thread to have its own connections so we don't have to contend for
them.

But in in-process databases, that is not possible: for an in-memory
sqlite database, if each executor thread has its own database they will
be effectively different databases with different data.

And that is the best case scenario: for an on-disk sqlite database, we
could end up with many different processes writing to the same file.

In here I am introducing the DbConnection struct, which captures that
and provides a method to derive a local connection from it. External
databases can open new connections at this time, whereas in-process
databases just clone their own state. The DbConnection struct implements
both Clone and Send, so it can be passed around to new executors.