Make ChiselStrike multithreaded
Each executor will spawn its own API server, and serve requests from it.
For this to work, the bound socket has to have SO_REUSEPORT. Because
hyper doesn't set that, we need to create our own TcpListener, set the
option, and then use a different method in Hyper to build a server from
an existing tcp socket.