Make CoordinatorChannel a bit easier to use
Now it has a send function that sends a command and waits for a
response.

While at it, move as much info as possible to CommandTrait, so that we
don't have to duplicate super long type signatures. It is not possible
to make Clone part of the CommandTrait since that would make it
object-unsafe:

https://doc.rust-lang.org/reference/items/traits.html#object-safety