accept typescript class definitions
We accept a typescript class in the type file.

There are constraints on what you can write, but the constraints are easy to document:

* nothing but classes in the file
* no method definitions inside the classes, including constructors
* attribute types must be base typescript types, or other types defined
  in the same file

Default values with literals are possible.

The parser is a bit ahead of chiseld: supports arrays, optionals, and
other things we haven't yet touched.

Supports decorators!