use a verb for apply, remove older verbs
As we are moving more and more towards a terraform-like workflow, it
doesn't make sense to keep the individual verbs on the RPC layer.

That doesn't matter much right now, but if we want to version the
endpoints together, plus data and policies, we have to receive them
as one versionable entity.

This PR is a stepping stone for that.