use pub(crate) in types.rs whenever possible
Unfortunately we can't just pattern match all uses, because some of
them, at least for now, do need to be pub. They seem to be the ones that
are eventually used either from typescript or options.

We'll have to convert them as we see them.

The advantage is that if a pub(crate) function is no longer referenced
anywhere, cargo will tell us about it.