add a version parameter to apply
Far away from ready, but enough to map where the issues are.

Allows you to do :

  chisel apply --version v0

And then:

  chisel apply --version v1

And now you have two branches that are completely independent, with
their own data and routes.

The use case for that is when the developer creates a new git branch
and wants to test endpoints there, in isolation.

An alternative design would be to leave chiselstrike as is, and for
a new branch we spawn a new chiseld fleet. However this will make it
harder to migrate data later, when the branch is merged.

For this design to work, we'll have to persist more state into chiseld.
Right now, we assume we'll just `chisel apply` after a restart, for
example. However, now that there are many endpoints in many versions,
you may not have access to the old versions anymore so you can't
recreate them.