Use only Swatinem/rust-cache@v1
Its documentation doesn't mention it, but the log of rust-cache shows

 Restoring paths:
    /home/runner/.cargo/bin
    /home/runner/.cargo/.crates2.json
    /home/runner/.cargo/.crates.toml
    /home/runner/.cargo/git
    /home/runner/.cargo/registry/cache
    /home/runner/.cargo/registry/index
    ./target

So the only missing file is ~/.cargo/.package-cache and it looks like
it is empty.

The real problem with the long builds seems to be:

 Using keys:
    v0-rust-build-1.59.0-nightly-x86_64-unknown-linux-gnu-48a5999fceee-68ae6e162e9a9a8d3be8
    v0-rust-build-1.59.0-nightly-x86_64-unknown-linux-gnu-48a5999fceee
 No cache found.

Which we have to debug, but setting up a second competing cache
doesn't seem to be the way to do it.