Access type fields through accessors.
We usually access name and backing_table directly, but once we introduce
versioned endpoints we'll have to mediate through versions. The backing
table, in particular, will be different for types that have the same
name but different versions. Also in preparation for the fact that we
will want to rename versions, the backing table name is generated
randomly.

One complication is that we currently test for equality by matching
ty1.backing_table == ty2.backing_table. But when we create a type object
and want to look it up, we don't yet know if the backing table will
match. To fix that we notice that, temporarily, we can't have types
with repeated names anyway so we just remove that part of the test.
When we introduce the api_version field, we will require that to match