use the Descriptor pattern to build fields
It worked quite well for the ObjectType, that has similar requirements.
This has the added nice immediate bonus of moving persisted name logic
into the type accessors methods. We do that to make the name field in
the database unique, but need to make sure it is parsed back before the
field is used.