Add a second, local-only web server
This will be used to support our K8s deployment. Liveness, readiness and
status checks are implemented (in their most simple forms for now)

To avoid having to worry about those being a source of DDoS, this is a
separate server that listens to 127.0.0.1 only. That's enough for k8s.