persist and load policies
Currently if we restart chiseld, the endpoints come back but the
policies don't. That's not a good match for the model we're evolving to.

The policies table has to be keyed by something. When we have versions,
we will key it by version. Instead of trying to come up with a temporary
schema for that table, we'll just use a temporary key.