add a newline to the example endpoint
lit is line-based, so not having the newline here is making it hard
to test lots of find in succession. So far we didn't hit this, but
I am hitting random errors in a new test of mine that seem to happen
due to line buffering, and this fixes it.