improve user experience when reading files
It took me a while now to figure out why chisel was returning simply
"    stream did not contain valid UTF-8"

The reason is because I had temporary files inside the endpoints
directory, but our test is broken. It tests for the dentry.path(),
which ends up being something like "/endpoints/.temporary.swp".

Instead, it should test or dentry.filename(), which is just the last
part. Also, we can't really accept anything that starts with ".", so
we can simplify the test.

In case we do fail for any other reason, I'm also improving the message
by listing which file failed.