Partial dependency update
I am trying to debug why "cargo install" fails. It is basically a
cargo update followed by cargo build, so this is a partial update
created by cargo update -p <packages> which still works.

This hopefully makes it easier to see which update breaks the
build. Sending it since an update is probably a good thing anyway.