add an internal Id field to ObjectTypes
Because it is internal it will not be visible to describe, and will
not be listed in the meta database.

It will be generated on the fly if one is not provided for writes.

For reads, users will now be able to say row.id and use it in their
endpoints.