bonus points: make chisel.store work for updates
This code now updates an object:

```
    let people = Person.rows();
    for await (let person of people) {
        if (person.country === country) {
            response += person.name + " " + person.id
            person.name = person.name + "FO"
            await Chisel.store('Person', person);
            response += "\n"
        }
    }
```

We do that by transforming the insert statement into an UPSERT.
If the user provides an ID, that will be used, and the query will be
expected to collide. If that is left empty, we generate one.

This works well because ids are now UUIDs, so even if the user provides
one, we can be certain it is safe to use it.