spill compilation errors to the client
Much, much better user experience. Some of this is temporary.
For example, we should be able to just run through the code without
having to call the entire define_endpoint. But for now this will do,
and we just add a BS endpoint at the __chiselstrike namespace

We have to adjust one of the tests for this to work. I will now argue
that the test as it was is not only wrong but dangerous:

What the test is testing for is:

 * successfully create an endpoint called test
 * test that /dev/test returns 200.
 * call chisel apply with some bad code now into test.ts
 * test that /dev/test returns 400

However chisel apply with bad code FAILS, so it should not touch what is
already working. We should always have validation steps happening before
any change is effectively attempted. This code introduces one such early
validation so the test starts failing, but the fact that it was passing
before is the scary part.