support version populate
This adds basic support for populate, however not implementing Lenses
yet since there is no universal agreement on the format.

Still, this should allow users to populate versions when types
are compatible, which is hopefully a big use case.

To speed up development I have used the lens implementation internally,
though, and just mapped the version into an empty lens.