Drop RoutePaths
Now that we have a PrefixMap with most of the logic, this was probably
too thin to be justified. It was also inflexible, always requiring
both the code and closure to be present, which is undesirable, since
api only needs the closure and the rest only the code.

As an example, server/src/query/meta/mod.rs no longer uses deno
directly.