Use consistent paths in apply_aux
We now do a first pass consuming apply_request.endpoints and adding
the version prefix. With this rust makes sure that nothing can use the
path without a prefix.

One advantage, in my opinion, is that errors now show the full
path. Another is that I don't get confused and try to use /foo when
/version/foo is needed.