add the type back into the Chisel namespace and provide a .d.ts file
This is because as previously explained, the type `Person` is an
instance of `Person`, but the linter expects something that returns
an async iterator here.

We toyed with the idea of calling it `person`, but because Typescript
is indeed case sensitive, we're just asking for problems if the user
defines for some reason - as stupid as it may be - both Person and
person.

One minor annoyance is that VScode doesn't auto-save the files (at
least not by default), and this is only generated if when you chisel
dev. But for now there isn't much we can do about it.