don't change the object (for real) when taking
Pekka noticed that passing this.inner to a new Table object in take
still modifies the original object because this is not a deep copy.

Unfortunately TypeScript doesn't have any builtin clone function,
so we just have to do it manually.

Adds a test to make sure we are respecting the original object.

Fixes #463