rename Table and table for api.ts
Rafael has correctly expressed in the past that we are seeing this being
named many things all over.

We all seem to agree that "Table" is problematic because users could
think of SQL tables.

So rename this to `ChiselIterator`, which is the same user-visible
name we are using in the dts files.