Make API usable without an async mutex
It is illegal to .await with the runtime held (RcMut), but we end up
doing that to get the API Service, since it is wrapped in a mutex.

It is wrapped in a Mutex, because is lifetime is managed by an Arc,
since Hyper expects it to be Send.

We can internalize the mutex and use a Sync Mutex instead. Because
we know this is TPC, we know that this mutex will never contend.
The alternative to the Mutex is to interior mutability with a RefCell
+ an unsafe forcing of Send. That actually works well, but it's
less future proof and would crash badly if someone ever uses this
outside of the TPC context by accident.