ChiselStrike npm package
This adds a ChiselStrike npm package into the tree.

The package consists of the following:

  - api.ts
  - chisel and chiseld binaries

The api.ts and lib_deno_core.d.ts are symlinked to the npm/src directory
to avoid churn, but at the same time, make sure the files are visible
when someone installs the package.

With this package, you can now do (once published):

  npm i @chiselstrike/chiselstrike

and you will have the latest version of the ChiselStrike API available.

At development time, you can always install the local version by first
building the package in the 'npm' directory:

  npm run build

and then installing from local filesystem with:

  npm i <chiselstrike.git>/npm

Please note that there are still work to be done to integrate this with
"chisel dev". First of all, we need to either generate package.json with
chiselstrike dependency in "chisel init" or make the chiselstrike
package install generate a Chisel.toml.

We also need to make "chisel" and "chiseld" deal with imports and
exports, but there is a pull request by Glauber to do that:

https://github.com/chiselstrike/chiselstrike/pull/512