materialize user code in the filesystem
Bringing in modules is really hard without this. Now that the
application is brought all at once in a version, we can have a temporary
directory that holds all modules.

api.ts and chisel.js are kept at the top level, and then we create
directories for the api_versions.

The code itself is still kept at the meta database, which allow us to
keep all the many versions in flight. These directories are only needed
when we are compiling the code and resolving modules, not when the code
is executing.

The approach is similar to what Rafael is doing, except that I have
a big directory created at program startup time and a directory per
project, instead of creating individual temporary files.