use tsc compiler in the system to check user files
If there isn't a tsc compiler available, we can still accept the files,
so we make an optional run. However to go the extra mile, we do specify
tsc as a runtime dependency for our npm package now.

If there is a tsc compiler available, then it passing becomes mandatory.
This way we can make sure all the tests are correct. In particular, for
our tests, we don't require tsc checking for simplicity, but could make
them stricter later.