accept arbitrary code in class definition
Now that we are moving most of the compilation to the client (and soon
all), there is no reason anymore to restrict the user to define
functions in the class definition.

In particular, the goal of this is to allow us to move from
Chisel.save() to person.save(): because in Typescript the only way to
set non-default properties is through a constructor, the user has to be
able to provide at least the constructor.