Rename 'types' directory to 'models'
One complaint we have heard is that the directory name 'types' is
confusing and ambiguous because developers are expected to put
TypeScript classes (not types) there.

But what should we call the directory? As the directory contains
entities, it's tempting to call it that. However, in common terminology,
entities are just the building block of something larger: application
domain *models*. IOW, a model is a collection of entity classes, and an
application can consist of multiple models.

And just to sanity-check that I did not invent this terminology myself,
examples include Microsoft's Entity Framework, Prisma, and Sequelize:

https://www.learnentityframeworkcore.com/model

https://www.prisma.io/docs/concepts/components/prisma-schema/data-model

https://sequelize.org/v7/manual/model-basics.html

Therefore, rename the directory to 'models'.o

Please note that we still retain the `ChiselEntity` base class as-is,
but we can debate if that ought to be renamed to `Model` as well.