server/chisel.ts: Entity class methods
Dejan found an issue where the save() method is missing if you query an
entity:

  import { Chisel } from "@chiselstrike/chiselstrike";
  import { User } from "../types/user";

  export default async function (req: Request): Promise<Response> {
      const user = new User();
      user.username = 'penberg';
      await user.save();
      const anotherUser = await User.findOne({ username: 'penberg'});
      await anotherUser.save();
      return Chisel.json(user);
  }
  penberg@vonneumann example % curl -d '{"username": "penberg"}' localhost:8080/dev/hello
  Uncaught TypeError: anotherUser.save is not a function
      at file:///var/folders/0n/v8sb64557vx78c6z4rqb8st40000gn/T/chiselstrike.uF9nMpoxNmoE//dev/hello.ts?ver=33:12:22

The problem is that we return entities from Deno ops as JSON, which has
no connection to the TypeScript class, and, therefore, all methods are
erased.

Let's fix the issue by explictly instantiating the correct type in the
`next()` function of ChiselIterator.

However, for join() and select(), we still retain the old behavior
because it's arguably the correct thing to do: the _type_ of a join() or
select() is not the entity, but a union of two types or a subset of a
type, respectively.

Fixes #549