Add type imports
It used to be possible to not include the source code of the type
definitions in what we send to chiseld, since they were used only for
types and not by javascript itself. That is no longer the case.

We currently use SWC, which accepts having both the copied code and
imports in the same compilation. When doing so SWC deletes the
imports.

With TSC, having both import and copied code produces errors like:

    Compilation failed:
    endpoints/.tmpk6ytmy.ts:4:14 - error TS2395: Individual declarations in merged declaration 'User' must be all exported or all local.

    4 export class User extends ChiselEntity {
                   ~~~~
    endpoints/.tmpk6ytmy.ts:7:10 - error TS2395: Individual declarations in merged declaration 'User' must be all exported or all local.

    7 import { User } from "../models/user.ts";
               ~~~~

As far as I can tell the options are:

1: Don't copy the code and implement support for sending multiple
   modules first. This is probably something we want to do, but it
   would keep TSC out of cli for longer.

2: Don't copy the code when feeding TSC and *then* copy it before
   sending to the endpoint. This is what I am currently trying.

Option 2 opens two suboptions. One is to synthesize a prefix of the
form

  import { foo } from "../foo.ts";
  import { bar } from "../bar.ts";

The other is to require user code to have those imports.

If I managed to follow the current changes, we do want user code to
include the imports. Producing the above prefix is also non trivial as
we currently don't keep track of which file defines which type.