Use TSC on the client
This is a fairly big hack. We first compile the real endpoint code
with TSC. This has the nice impact of producing errors pointing to the
correct file and line number.

Once the code has been compiled, we concatenate the type definition
code, as we do on main. The resulting code is passed to SWC to delete
imports that the server cannot handle until we are sending multiple
modules.