Use types= to refer to lib.deno_core.d.ts
I confess to not fully understand all the differences between path=
and types=. It looks like types= is intended for .d.ts files, so this
is probably correct. More importantly, this is the only way I could
figure out to keep the standalone tsc working and switch cli to use
tsc_compile.