server/chisel.ts: Implement ChiselIterator.skip(N)
This implements a ChiselIterator.skip(N) method. The specified N will be
turned into the following SQL query:

 SELECT * FROM ... LIMIT M OFFSET N

Please note that it's invalid to call skip() without take() because
databases such as SQLite require LIMIT to go with OFFSET. There's a
"LIMIT ALL" in PostgreSQL, but MySQL, for example, is similar to SQLite.

Please note that the ordering of skip() and limit() matters. We take the
semantics as if we were really traversing an iterator. Therefore,

  - skip(2).take(4) => LIMIT 4 OFFSET 2
  - take(4).skip(2) => LIMIT 2 OFFSET 2
  - skip(2).take(4).skip(1) => LIMIT 3 SKIP 3

and so on.