Use tsc in the server
For now tsc is added at tsc_compile and used only in the server. Using
it in the client with what is currently in main requires a bit more
work.

In a followup PR I will move the chisel.ts compilation to the server
build time.

This uses done in a submodule at tag v1.17.3, which corresponds to the
version we are using. There is a submodule update in build.rs.