cli: Ignore file metadata changes in 'chisel dev'
Currently, 'chisel dev' happly applies multiple times when a file is
saved in VS Code, for example. Fix the code to only apply when file
contents change, but ignore file metadata changes.

After this change, "apply" (for the whole project) is run exactly once
per file saved. As a future improvement, we should teach "chisel apply"
to check if file contents are identical, and make it a no-op.