don't wrap deno in an rcmut
We need to await with it held, which violates the RcMut rules.
We can't play the same tricks we did with our Runtime structure, because
Deno expects its methods to be called with &mut self and not &self.

So use an async mutex instead. The async mutex is fine, as long as we
don't do reentrancy. It effectively limits the concurrency to
nr_threads, or 1 if there is a single thread.

It has the side-effect of effectively guaranteeing a transaction
boundary per endpoint in the case of a single thread.

The test attached panics without the patch

Fixes #650