deduplicate module imports before writing types file into endpoint
Since Pekka ranked this as high priority, I took a stab at a quick hack
to fix the issue of having multiple files with models. The approach is
simple and should cause minimum conflicts.

It essentially reads all the types first and deduplicates imports. It
doesn't try to parse the contents at all, and if you end up importing
twice, but with different syntax (there are many valid import syntaxes
for typescript), it won't work. It will also not work if you happen to
import ChiselEntity in the endpoint code as well.

However, the path of VSCode autocompleting will work, because the
string will always be the same. And ChiselEntity shouldn't really be
declared in endpoint code, because we don't accept models there.

So hopefully this is an acceptable solution that can make our release
a bit better.

Refs: #645