Avoid must_not_suspend warnings
Also, don't hold locks in futures. This requires a bad hack, but I am
still tempted by it, since right now we can only handle one request at
a time. This is both slow and a very good place for bugs to hide.

This brings

$ perf stat -r 5 ../target/debug/deps/integration_tests-...--test multiple.lit

from

4.62859 +- 0.00437 seconds time elapsed  ( +-  0.09% )

to

2.61810 +- 0.00813 seconds time elapsed  ( +-  0.31% )

By allowing both requests to overlap.