Adopt webpack for node-style resolution
This PR adds a webpack step before sending the endpoints to
ChiselStrike. Although there is more work to be done (we could
use chunks to deduplicate code in common for various endpoints, and
stuff that depends on the node global variables don't yet work), this
already elegantly fixes the issue of repeated imports in multiple models
files.

It also should completely free Rafael to move on with the tsc work, as
it demarks the node / deno issue in a very clear manner. A project
shouldn't be able to mix them, and everything that wants to behave like
node has to go through webpack.

Which module style we'll follow is now determined in Chisel.toml.
Users that create a project through npm, get node-style module
resolution by default. Users who use `chisel init` get deno-style.
I also propose we document this different clearly.

In terms of executing webpack, we need its cli because the alternative
is integrating with node, and the whole chicken and egg issue we have is
lack of node. We also need the ts-loader webpack's plugin so we can
write typescript code.