improve build API
The Object.assign() API, that we use as a building block to build(),
can take many JSON objects. We can only take one, so in situations where
there are more than one you have to externally compose it to one first
(using Object.assign...), and then pass it to build.

Let's improve the API by adopting the same as Object.assign. The
semantics are well-defined and users are used to it.

Because a single object is a special case of many objects, this is
a compatible change.