manage secrets
Allow secrets to be used from the endpoint code. Secrets are essentially
a json file from an URL, which allow us to have more complex secrets
than just key-values.

Another advantage of being a file, and not an environment variable,
is that we can provide hot-reload.

Local dev users can add a .env file to their projects, and test locally.
For k8s, we'll probe an URL where a ChiselStrike server will provide
access to the encrypted secrets. The private key will be mounted to the
container and ChiselStrike itself will, in-memory, decrypt it.

Fixes #693