Restore endpoint `__chiselstrike/auth/user`
This was dropped by 166be68, because the user can make an endpoint
that returns Chisel.loggedInUser().  But we don't want to depend on
that in the frontend package; we want frontends to conveniently say
`client.user`.

Fixes #710.

Signed-off-by: Dejan Mircevski <dejan@chiselstrike.com>
(cherry picked from commit 0d2b416fc1725ffb8252532a96d105f9b4f35859)