Add lib.dom.asynciterable.d.ts to the default libraries
This is a reduced testcase of the issue in #713

The story behind this is long, but the summary is that the
lib.dom.d.ts in TSC is behind the standard and lacks async iteration
in ReadableStream. The code in deno both provides and uses async
iteration, so trying to use lib.dom.d.ts instead of deno's own d.ts
files causes the code using it to fail to compile.

We could use deno's own d.ts files, like deno does, except we require
being able to run the standalone TSC, and it knows nothing about deno.

Fortunately deno has a lib.dom.asynciterable.d.ts that can be used to
patch up lib.dom.d.ts from TSC. We can use that as long as we don't
need to run the stand alone TSC on a file that requires async
iteration on ReadableStream.