Don't emit declarations by default
For reasons that are probably best left ununderstood, there are files
that TSC can process, but not produce declarations for.

Deno doesn't ask TSC to produces declarations when it just wants to
run a .ts file, so it could handle files that we could not.

This patch adds a emit_declarations option and makes it false by
default.

This fixes another testcase reduced from #713.