Handle exceptions in compile
This produces a diagnostics if an exception gets to the top level
compile. At some point we will want to handle them lower in the call
stack and produce plain errors, but having a general handling of
missing cases is probably a good thing.

This is also fairly handy for current development.