Return a dummy module when one is missing
Deno does a first pass with SWC to download all the modules. It is in
this pass that it reports broken imports.

Given the first pass, deno knows that if TSC asks for a module it
doesn't have, it is a dynamic import and it feeds TSC a dummy module.

We don't yet have a first pass, but we want to avoid errors at dynamic
imports, so also feed TSC dummy modules when we can't resolve them.

This regresses our error messages until the deno code for using SWC is
refactored and shared, but is another step in fixing #713.