fix multiple endpoint case in node mode
When in node mode, we were using the same temporary output file for
all endpoints. In practice, for cases with multiple endpoints we just
ended up with N copies of one of them.

We do have a test for that, but it only exercises the deno path.
The test is enhanced to test for node mode too. For that to happen
we need esbuild to be included in our package.json