make it easier to write node-mode tests.
Tests with the extension `.node` instead of `.lit` will use
create-chiselstrike-app to create the package and use node-style
resolution through esbuild.

This should make it easier for us to make sure that this path - which
is what most users are using - will keep working.

An initial test is added with a follow up recently requested by Rafael,
to test the handlebars import.