build packages at build time
Our tests are depending more and more on some of our npm packages,
but they are not built with the project, mostly because they are node/ts
entities.

Doing those things at test time, we are then either relying on ordering
that the tests execute or duplicating work.

Move this to the build step so we are sure they are always present.
We create an empty project, since the packages directory is all
typescript, just so we can execute it's build.rs from `cargo build`