log context of failed queries
Sometimes query fails, and the error messages give us no indication
which one failed. We should log the context of which query was being
executed when it failed.

This is done through macros because sqlx SUCKS and I couldn't get the
lifetimes to work.

I've attempted doing this in the past but it wasn't passing tests,
so this is a fresh take that relies less on macro magic and expects
a query to be fully formed by the time it reaches the macro (and not
an expression that computes to a query). Seems to work.