allow uniqueness to work with postgres
A couple of issues in the uniqueness patch when working on postgres.

The first, and hardest, is that in sqlite we can probe a table to see
if exists. A select failing doesn't error out the transaction, so we
could do the trick of checking the error on a table we knew had to
exist.

In Postgres, failing a read does fail the transaction so we can't do
that. Instead, we unfortunately have to go and check if those things
exist. The way to check is different for postgres and sqlite. We have
no choice but to just do it the way those databases prefer.

Also, for Postgres 'chiselstrike' and "chiselstrike" are different
things, so in getting the version we have to do 'chiselstrike'.

Aside from that, we also have the issue with the upserts that we
recently found